post and category sections are completed 100%

+++ app/Http/Controllers/AdminUsersController.php
@@ -133,9 +133,12 @@ class AdminUsersController extends Controller
     {
         //
         $user = User::findOrFail($id);
-        unlink(public_path() . $user->photo->file);
+        if ($user->photo) {
+            unlink(public_path() . $user->photo->file);
+        }
+        
         $user->delete();
-        // I added this to delete the photo
+        // I added this to delete the photo in the database
         $user->photo()->delete();
 
         Session::flash('delete_user', 'User deleted');
